+++ src/main/java/com/recipe/app/src/common/client/naver/dto/NaverAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.common.client.dto;
+package com.recipe.app.src.common.client.naver.dto;
 
 import com.recipe.app.src.user.application.dto.UserLoginRequest;
 import lombok.Getter;
